pyproject.toml
[tool.poetry]
name = "email-spam-detection"
version = "0.1.0"
description = ""
authors = ["Jakub Ciszek <jakubciszek64@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytorch-lightning = "^1.5.1"
wandb = "^0.12.6"
dvc = {version = "^2.8.3", extras = ["gdrive"]}
pandas = "^1.3.4"
tqdm = "^4.62.3"
transformers = {extras = ["sentencepiece"], version = "^4.12.5"}
seaborn = "^0.11.2"
scikit-learn = "^1.0.1"
torchmetrics = "^0.6.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
ipykernel = "^6.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
